+++ TUPScheduling/templates/account/login_old.html
@@ -1,0 +1,86 @@
+{% extends "base.html" %}
+
+{% load i18n %}
+{% load account socialaccount static wagtailimages_tags %}
+
+{% block head_title %}{% trans "Sign In" %}{% endblock %}
+
+{% block content %}
+
+
+{% get_providers as socialaccount_providers %}
+
+
+<section class="flex flex-row h-screen mx-auto">
+
+    
+  <div class="flex flex-row flex-1 ">
+      
+      <div class="flex flex-col w-10 ">
+          <div class="flex items-end justify-end flex-1 mb-2 -mr-3">
+              <h1 class='font-sans text-xl' style="writing-mode: vertical-rl; transform: rotate(-180deg); color: #575757;">Ermita Manila</h1>
+          </div>
+          <div class="flex-1 h-1/2" style="border-right: solid 4px #C7C9CD;">
+          </div>
+      </div>
+      
+      <form style="color: #960404;" class="flex flex-col items-center justify-center flex-1 px-4 login"  method="POST" action="{% url 'account_login' %}">
+          <div>
+              <h1 class="font-sans text-2xl font-bold text-center">TECHNOLOGICAL UNIVERSITY</h1>
+              <h1 class="font-sans text-2xl font-bold text-center">OF THE PHILIPPINES</h1>
+  
+              {% csrf_token %}
+              
+              <div class="w-full mt-16 space-y-4">
+                  <div class="flex flex-row"  style="border: solid 4px #C4C5C8;">
+                      <div class="px-4 " style="border-right: solid 4px #C4C5C8; padding-top: 4px; padding-bottom: 4px;">
+                          <img src='{% static "img/login/user.webp" %}' style="width: 2rem;" alt="User Icon">
+                      </div>
+                      <input class="w-full px-4 text-black" name='login' required id='id_login' autocomplete="email" type='text'>
+                  </div>
+  
+                  <div class="flex flex-row"  style="border: solid 4px #C4C5C8;">
+                      <div class="px-4 " style="border-right: solid 4px #C4C5C8; padding-top: 4px; padding-bottom: 4px;">
+                          <img src='{% static "img/login/key.webp" %}' style="width: 2rem;" alt='Key Icon'>
+                      </div>
+                      <input class="w-full px-4 text-black" name='password' required id='id_password' autocomplete="password" type='password'>
+                  </div>
+              </div>
+
+              <div class="flex flex-row items-center justify-between py-10">
+                  <a href="{% url 'account_reset_password' %}" style='color:#575757;' class='font-sans text-lg font-medium' >Forgot password?</a>
+                  <button  type="submit" style="background-color: #960404; color: white;" class="px-8 py-2" aria-label="Login Button">LOGIN</button>
+              </div>
+
+              <div class="flex flex-row items-center justify-center pb-10">
+                  <hr class="flex-1" style="border-top: 3.5px dotted #898B8F;">
+                  <h1 class="px-2 text-lg font-bold" style="color: #575757;">OR</h1>
+                  <hr class="flex-1" style="border-top: 3.5px dotted #898B8F;">
+              </div>
+              
+              
+              <button style="background-color: #960404; color: white;" class="flex flex-row items-center w-full pr-4 space-x-4" aria-label="Login Gmail Button">
+                  <div style="background-color: #A42F2F;" class="px-4 py-2">
+                      <img src='{% static "img/login/gmail.webp" %}' style="width: 2rem;" alt='Gmail Icon'>
+                  </div>
+                  <h1 class="font-sans font-bold text-md">LOGIN TO TUP GMAIL</h1>
+              </button>
+          </div>
+          
+      </form>
+
+      <div class="w-10 h-1/2" style="border-left: solid 4px #C7C9CD;">
+      </div>
+  </div>
+
+  
+
+  <div style="flex: 1.3; overflow: hidden;">
+      <img src='{% static "img/login/tup-bg.webp" %}' class="w-full" alt='Right Background Picture' style="object-fit: contain;"/>
+  </div>
+
+
+</section>
+{% endblock %}
+
+
