+++ TUPScheduling/templates/room_schedule/room_schedule.html
@@ -1,0 +1,144 @@
+
+{% extends 'base.html' %}
+
+{% load static wagtailcore_tags %}
+
+
+{% block header %}
+<!--Header-->
+<header>
+    <div class="p-10 font-bold text-center place-self-center">
+        <a href="#">
+            <img class="mx-auto mb-2" src="{% static 'img/icons/tuplogo.webp' %}" alt="tup logo">
+            <h1>TECHNOLOGICAL UNIVERSITY</h1>
+            <h1>OF THE PHILIPPINES</h1>
+        </a>
+    </div>
+</header>
+
+<!--Navbar-->
+<nav class="bg-red-700 ">
+    <div class="flex justify-end text-white lg:hidden">
+      <button class="p-4 mobile-menu-button focus:outline-none focus:bg-red-600">
+        <svg class="w-5 h-5" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 6h16M4 12h16M4 18h7" />
+        </svg>
+      </button>
+    </div>
+  
+    <div class="flex justify-center">
+      <div class="absolute inset-y-0 left-0 z-10 px-2 space-y-6 transition duration-150 ease-in-out transform -translate-x-full bg-red-700 sidebar lg:relative lg:translate-x-0">
+        <div class="w-full lg:flex">
+          <div class="pb-4 lg:flex md:pb-0">
+            <ul class="pb-2 text-white md:pb-0 lg:flex ">
+              <li class="mr-20"><a class="block px-0 py-2 pt-4 border-b-2 border-transparent lg:px-3 hover:border-white" href="/class-schedule"><span>CLASS SCHEDULE</span></a></li>
+              <li class="mr-20"><a class="block px-0 py-2 pt-4 border-b-2 border-transparent lg:px-3 hover:border-white" href="/instructor-schedule"><span>INSTRUCTOR SCHEDULE</span></a></li>
+              <li><a class="block px-0 py-2 pt-4 border-b-2 border-transparent border-white lg:px-3" href="/room-schedule"><span>ROOM SCHEDULE</span></a></li>
+            </ul>
+          </div>
+        </div>
+      </div>
+    </div>
+</nav>
+{% endblock %}
+
+{% block content %}
+<!--Rooms Image Grid-->
+<section>
+    <div class="mx-12 mt-20 md:mx-40">
+      <div class="w-56 border-l-8 border-red-700">
+        <h2 class="py-2 ml-2 text-3xl font-bold uppercase">rooms</h2>
+      </div>
+    </div>
+
+    <div class="mx-12 mt-12 mb-20 md:mx-40">
+      <div class="grid grid-cols-1 gap-6 p-8 bg-gray-100 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-5 rounded-xl">
+        <a href="laboratory room.html">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image2.png' %}" alt="lab">
+            <h2 class="pt-2 font-bold text-center text-red-700">LABORATORY 1</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image2.png' %}" alt="lab">
+            <h2 class="pt-2 font-bold text-center text-red-700">LABORATORY 2</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image2.png' %}" alt="lab">
+            <h2 class="pt-2 font-bold text-center text-red-700">LABORATORY 3</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image4.png' %}" alt="lecture">
+            <h2 class="pt-2 font-bold text-center text-red-700">ROOM 311</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image4.png' %}" alt="lecture">
+            <h2 class="pt-2 font-bold text-center text-red-700">ROOM 312</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image4.png' %}" alt="lecture">
+            <h2 class="pt-2 font-bold text-center text-red-700">ROOM 313</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image4.png' %}" alt="lecture">
+            <h2 class="pt-2 font-bold text-center text-red-700">ROOM 314</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image4.png' %}" alt="lecture">
+            <h2 class="pt-2 font-bold text-center text-red-700">ROOM 315</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image4.png' %}" alt="lecture">
+            <h2 class="pt-2 font-bold text-center text-red-700">ROOM 316</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+        <a href="#">
+          <div class="bg-white rounded-b-xl">
+            <img class="object-cover w-full" src="{% static 'img/image4.png' %}" alt="lecture">
+            <h2 class="pt-2 font-bold text-center text-red-700">ROOM 317</h2>
+            <p class="pb-4 font-thin text-center">COS Building - 3rd Floor</p>
+          </div>
+        </a>
+      </div>
+    </div>
+</section>
+{% endblock %}
+
+{% block scripts %}
+<!--JS Scripts-->
+<script>
+
+  //Sidebar
+  const btn = document.querySelector(".mobile-menu-button");
+  const sidebar = document.querySelector(".sidebar");
+  
+  
+  btn.addEventListener("click", () => {
+    sidebar.classList.toggle("-translate-x-full");
+  });
+</script>
+{% endblock %}
