+++ administrator/__init__.py
@@ -1,0 +1,22 @@
+_DAY = [
+    'Monday',
+    'Tuesday',
+    'Wednesday',
+    'Thursday',
+    'Friday',
+    'Saturday'
+]
+
+_TIME = [
+    '7:00 - 8:00',
+    '8:00 - 9:00',
+    '9:00 - 10:00',
+    '10:00 - 11:00',
+    '11:00 - 12:00',
+    '12:00 - 1:00',
+    '1:00 - 2:00',
+    '3:00 - 4:00',
+    '4:00 - 5:00',
+    '5:00 - 6:00',
+    '6:00 - 7:00',
+]
