+++ instructor_schedule/models.py
@@ -1,0 +1,6 @@
+from django.db import models
+from wagtail.core.models import Page
+
+
+class InstructorSchedule(Page):
+    parent_page_types = ["home.HomePage"]
